fix(mdx): convert remark-images-to-component plugin to a rehype plugin
Image related rehype plugins was not working because the MDX
integration used a remark plugin. Converting it to a rehype plugin
allow us to keep the same behavior but after any user's plugins run.

Fix #10643
