+++ .changeset/sweet-goats-own.md
@@ -1,0 +1,5 @@
+---
+"@astrojs/mdx": major
+---
+
+Replace remark-images-to-component with rehype-images-to-component to let users use additional rehype plugins for images
